feat: integrate all progress

+++ database/migrations/2026_04_24_042241_create_categories_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('icon');
             $table->string('name');
-            $table->decimal('monthly_budget');
+            $table->unsignedBigInteger('monthly_budget');
             $table->timestamps();
         });
     }
